updated nav, header, crud for charities, and myUser on all endpointts

+++ src/main/java/com/javaproject/javaprojectthree/service/impl/CharityServiceImpl.java
@@ -46,43 +46,49 @@ public class CharityServiceImpl implements CharityService {
 
     @Override
     public Charity createCharity(String title, String description, double goal, double totalReceived, Boolean verified, String pictureURL) {
+        if (JavaProjectThreeApplication.myUserDetails != null) {
+            if (!charityRepository.existsByTitle(title)) {
 
-        if (!charityRepository.existsByTitle(title)) {
-
-            Charity newCharity = new Charity();
-            newCharity.setTitle(title);
-            newCharity.setDescription(description);
-            newCharity.setGoal(goal);
-            newCharity.setTotalReceived(totalReceived);
-            newCharity.setVerified(false);
-            newCharity.setPictureURL(pictureURL);
-            return charityRepository.save(newCharity);
+                Charity newCharity = new Charity();
+                newCharity.setTitle(title);
+                newCharity.setDescription(description);
+                newCharity.setGoal(goal);
+                newCharity.setTotalReceived(totalReceived);
+                newCharity.setVerified(false);
+                newCharity.setPictureURL(pictureURL);
+                return charityRepository.save(newCharity);
+            } else {
+                throw new InformationExistException("Charity with the name " +
+                        title + " already exists");
+            }
         } else {
-            throw new InformationExistException("Charity with the name " +
-                    title + " already exists");
+            throw new InformationNotFoundException("Must be logged in to perform this action");
         }
     }
-
     @Override
     public ResponseEntity<?> createCharity(RegisterForm registerForm) {
+        if(JavaProjectThreeApplication.myUserDetails != null) {
+            if (!charityRepository.existsByTitle(registerForm.getTitle())) {
 
-        if (!charityRepository.existsByTitle(registerForm.getTitle())) {
-
-            Charity newCharity = new Charity();
-            newCharity.setTitle(registerForm.getTitle());
-            newCharity.setDescription(registerForm.getDescription());
-            newCharity.setGoal(registerForm.getGoal());
-            newCharity.setTotalReceived(0);
-            newCharity.setVerified(false);
-            newCharity.setPictureURL(registerForm.getPictureURL());
-            newCharity.setUser(
-                    userRepository.findUserByUserName(registerForm.getUsername())
-            );
-            return ResponseEntity.ok(charityRepository.save(newCharity));
-        } else {
-            throw new InformationExistException("Charity with the name " +
-                    registerForm.getTitle() + " already exists");
+                Charity newCharity = new Charity();
+                newCharity.setTitle(registerForm.getTitle());
+                newCharity.setDescription(registerForm.getDescription());
+                newCharity.setGoal(registerForm.getGoal());
+                newCharity.setTotalReceived(0);
+                newCharity.setVerified(false);
+                newCharity.setPictureURL(registerForm.getPictureURL());
+                newCharity.setUser(
+                        userRepository.findUserByUserName(registerForm.getUsername())
+                );
+                return ResponseEntity.ok(charityRepository.save(newCharity));
+            } else {
+                throw new InformationExistException("Charity with the name " +
+                        registerForm.getTitle() + " already exists");
+            }
         }
+        else {
+            throw new InformationNotFoundException("Must be logged in to perform this action");
+            }
     }
 
     @Override
@@ -102,13 +108,17 @@ public class CharityServiceImpl implements CharityService {
         return charityRepository.findCharityById(id);
     }
 
-    public void deleteCharity(Long charityId){
-        System.out.println("Service calling deleteCharity ==>");
-        Optional<Charity> charity = charityRepository.findById(charityId);
-        if (charity.isPresent()) {
-            charityRepository.deleteById(charityId);
+    public void deleteCharity(Long charityId) {
+        if (JavaProjectThreeApplication.myUserDetails != null) {
+            System.out.println("Service calling deleteCharity ==>");
+            Optional<Charity> charity = charityRepository.findById(charityId);
+            if (charity.isPresent()) {
+                charityRepository.deleteById(charityId);
+            } else {
+                throw new InformationNotFoundException("Charity with id " + charityId + " not found");
+            }
         } else {
-            throw new InformationNotFoundException("Charity with id " + charityId + " not found");
+            throw new InformationNotFoundException("Must be logged in to perform this action");
         }
     }
 
@@ -162,14 +172,17 @@ public class CharityServiceImpl implements CharityService {
     @Override
     public void update(Charity charity)
             throws InformationNotFoundException, InformationExistException {
-        if (!StringUtils.isEmpty(charity.getTitle())) {
-            if (!charityRepository.existsByTitle(charity.getTitle())) {
-                throw new InformationNotFoundException("Cannot find Contact with id: " + charity.getId());
+        if (JavaProjectThreeApplication.myUserDetails != null) {
+            if (!StringUtils.isEmpty(charity.getTitle())) {
+                if (!charityRepository.existsByTitle(charity.getTitle())) {
+                    throw new InformationNotFoundException("Cannot find Contact with id: " + charity.getId());
+                }
+                charityRepository.save(charity);
+            } else {
+                throw new InformationNotFoundException("Failed to save contact");
             }
-            charityRepository.save(charity);
-        }
-        else {
-            throw new InformationNotFoundException("Failed to save contact");
+        } else {
+            throw new InformationNotFoundException("Must be logged in to perform this action");
         }
     }
